+++ Nasmo/src/app/application.cpp
@@ -19,7 +19,7 @@ void nsm::Application::run() {
     this->intermoduleDataTransfer();
 
     do { // Main loop
-        const f32 ts = mDeltaTime.update();
+        const f64 ts = mDeltaTime.update();
 
         this->onUpdate(ts);
         this->handleEvents();
