+++ Nasmo/src/app/deltatime.cpp
@@ -1,8 +1,8 @@
 #include <nsm/app/deltatime.h>
 
-f32 nsm::DeltaTime::update() {
+f64 nsm::DeltaTime::update() {
     auto current = std::chrono::steady_clock::now();
-    mTimeStep = std::chrono::duration<f32>(current - mLastTime).count();
+    mTimeStep = std::chrono::duration<f64>(current - mLastTime).count();
     mLastTime = current;
 
     return mTimeStep;
