+++ Sandbox/src/entity/lightentity.cpp
@@ -27,7 +27,7 @@ public:
         this->addComponent<nsm::DrawableComponent>(mLightComponent);
     }
 
-    void onUpdate(const f32 timeStep) override {
+    void onUpdate(const f64 timeStep) override {
         static f32 posX = mTransform->getPosition().x, posY = mTransform->getPosition().y, posZ = mTransform->getPosition().z;
         static f32 colorR = 1.0f, colorG = 1.0f, colorB = 1.0f;
         static f32 intensity = mIntensity;
